ARCH-1: extract lint dispatch arm into commands::lint::run (iter-225)
